# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Trademark Journal Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Dashboard {
  id: number;
  /** @nullable */
  journalNo: string | null;
  /** @nullable */
  journalDate: string | null;
  keywordMatchMode?: string;
}

export interface DashboardInput {
  journalNo?: string;
  /** @nullable */
  journalDate?: string | null;
  keywordMatchMode?: string;
}

export interface Keyword {
  id: number;
  value: string;
}

export interface KeywordInput {
  value: string;
}

export interface TmNumber {
  id: number;
  value: string;
}

export interface TmNumberInput {
  value: string;
}

export interface JournalEntry {
  id: number;
  applicationNo: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  classNo?: string | null;
  /** @nullable */
  applicant?: string | null;
  /** @nullable */
  agent?: string | null;
  /** @nullable */
  filingDate?: string | null;
}

export interface JournalPage {
  entries: JournalEntry[];
  total: number;
  page?: number;
  limit?: number;
}

export type JournalImportEntriesItem = {
  applicationNo: string;
  title?: string;
  classNo?: string;
  applicant?: string;
  agent?: string;
  filingDate?: string;
};

export interface JournalImport {
  entries: JournalImportEntriesItem[];
}

export interface ImportResult {
  imported: number;
  skipped: number;
  message?: string;
}

export interface ReviewItem {
  id: number;
  journalNo: string;
  /** @nullable */
  journalDate?: string | null;
  applicationNo: string;
  /** @nullable */
  caseNo?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  classNo?: string | null;
  /** @nullable */
  applicant?: string | null;
  /** @nullable */
  agent?: string | null;
  /** @nullable */
  filingDate?: string | null;
  /** @nullable */
  matchedTerm?: string | null;
  /** @nullable */
  matchType?: string | null;
  status: string;
  /** @nullable */
  generatedDoc?: string | null;
  /** @nullable */
  generatedPdf?: string | null;
  /** @nullable */
  generatedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}

export interface ReviewItemUpdate {
  status?: string;
  notes?: string;
  generatedDoc?: string;
  generatedPdf?: string;
}

export interface TmRecord {
  id: number;
  journalNo: string;
  /** @nullable */
  journalDate?: string | null;
  /** @nullable */
  tmNo?: string | null;
  /** @nullable */
  caseNo?: string | null;
  /** @nullable */
  classNo?: string | null;
  /** @nullable */
  applicant?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  agent?: string | null;
  applicationNo: string;
  /** @nullable */
  filingDate?: string | null;
  /** @nullable */
  matchedTerm?: string | null;
  /** @nullable */
  matchType?: string | null;
  /** @nullable */
  remarks?: string | null;
  /** @nullable */
  docUrl?: string | null;
  /** @nullable */
  pdfUrl?: string | null;
  createdAt?: string;
}

export interface ScanResult {
  matched: number;
  skipped: number;
  total: number;
  journalNo?: string;
  message?: string;
}

export interface ActionResult {
  success: boolean;
  message: string;
  count?: number;
}

export interface ExportResult {
  success: boolean;
  message: string;
  files?: string[];
}

export type StatsReviewByStatusItem = {
  status: string;
  count: number;
};

export interface Stats {
  totalJournalEntries: number;
  totalReviewItems: number;
  totalRecords: number;
  pendingReview?: number;
  approvedForMerge?: number;
  generated?: number;
  keywordMatches?: number;
  tmMatches?: number;
  reviewByStatus: StatsReviewByStatusItem[];
}

export type GetJournalEntriesParams = {
page?: number;
limit?: number;
search?: string;
};

export type GetReviewItemsParams = {
status?: string;
journalNo?: string;
matchType?: string;
};

export type GetRecordsParams = {
journalNo?: string;
search?: string;
};