# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Trademark Journal Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface JournalEntry {
  id: number;
  applicationNo: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  classNo?: string | null;
  /** @nullable */
  applicant?: string | null;
  /** @nullable */
  agent?: string | null;
  /** @nullable */
  filingDate?: string | null;
}