# lib/api-zod/src/generated/types/reviewItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Trademark Journal Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewItem {
  id: number;
  journalNo: string;
  /** @nullable */
  journalDate?: string | null;
  applicationNo: string;
  /** @nullable */
  caseNo?: string | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  classNo?: string | null;
  /** @nullable */
  applicant?: string | null;
  /** @nullable */
  agent?: string | null;
  /** @nullable */
  filingDate?: string | null;
  /** @nullable */
  matchedTerm?: string | null;
  /** @nullable */
  matchType?: string | null;
  status: string;
  /** @nullable */
  generatedDoc?: string | null;
  /** @nullable */
  generatedPdf?: string | null;
  /** @nullable */
  generatedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}