# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Trademark Journal Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { StatsReviewByStatusItem } from './statsReviewByStatusItem';

export interface Stats {
  totalJournalEntries: number;
  totalReviewItems: number;
  totalRecords: number;
  pendingReview?: number;
  approvedForMerge?: number;
  generated?: number;
  keywordMatches?: number;
  tmMatches?: number;
  reviewByStatus: StatsReviewByStatusItem[];
}