# lib/api-zod/src/generated/types/tmNumber.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Trademark Journal Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface TmNumber {
  id: number;
  value: string;
}